Fixing bug due to 'keyish' field in PriceFixingsHistoryDataType
exchange needs to be a key field in order to see Libor like market data, but exchange isn't always available

So I've change the code to cease expecting all key fields to be available (at least when applying the override rule).

Perhaps a better fix would be to define a new type of field that behaves like a key when present.