FX rates do not appear to be updating QV
Enabled QlikView by default
Ensured default 'QlikViewSpotFXTask' is invalid as no cross-deployment default exists.
Added error logging when QlikView is notified but the latestSpotFXRates market data service method is not called within 2 seconds, this indicates that either QlikView is down or the task that was run is pointing to a different environment.