Meads got strange errors for LME trades when runing a report with All Closes and 02Jan2012.
This was because LIM returns 0 for all prices for this day, which results in a failed assertion,
which is then caught and re-thrown with a general no price found exception.

I have changed this so that the zero prices are filtered out early and if that leaves no prices a specific error is shown.

I could have filtered out all zero prices from LIM but maybe for some market somewhere zero is valid (eg spreads).

LIM is confusing with zeros. If you ask for prices in the range 02jan2012->03Jan2012 you only get the price for 03jan
but when you ask for a price for 30dec2011->03Jan2012 it returns the 0 price for 02jan2012.