Instruments dynamically resolve labels to Markets and Indices.
Author: Timur Alavidze <timur.alavidze@trafigura.com>
Author: Maxim Klivadenko <maxim.klivadenko@trafigura.com>
Author: Eugene Debrev <Eugene.Debrev@trafigura.com>

This change is meant to be the first step towards close-of-book and historical analysis functionality in Starling.  To be able to faithfully reconstruct past states of the system, Starling needs to have its data versioning extended to reference data: markets and indices, calendars, expiry rules.  This change brings dynamic reference data resolution---markets and indices in this chunk.  Other types of reference data and the corresponding versioned storages are to follow.

The key moving parts are as follows.

Instead of references to markets and indices, Instruments now contain *market labels*.  These are identifiers dynamically resolved to obtain a Market or an Index.  There are several types of market labels, all of them growing from the MarketLabel trait.  Most importantly, IdLabel refers to persistent markets and indices, a bunch of other types refer to indices which are created on the fly (futures contract, etc).  UnresolvedLabel is a special case---a reference by name---which will hopefully go soon.

To resolve a label to the actual Market or Index, a *resolution context* is required.  Conceptually, this context corresponds to a particular state of reference data.  This state needs to be specified in some reasonable way before Instruments can be evaluated or before some of the validations can be run, because these computations depend on information associated with markets and indices: UOMs, currency, lot sizes, rounding, etc.  So, quite a few methods have acquired an additional IndexResolver or Environment parameter; IndexResolver is a trait representing a resolution context, Environment extends this trait for convenience reasons.

For now persistence is not strongly affected, however, one of the next steps would be to use "internal Starling identifiers" to connect Instruments to Markets and Indices.  At the moment we write and read names; when a market/index name is read, we resolve reference data using the most recent version and extract a label---this is a temporary solution until we rework the persistence.

Certain changes to Starling's user-visible behaviour are introduced.

One of the major is related to the "expiryDay"---an Instrument declaring that past this date it has no position/risk, and Starling applies a filter to remove presumably "expired" trades from processing.  Some Instruments had their expiry days defined using calendars associated with markets and indices; this is now unavailable without a resolution context, and consequently the set of trades may be a bit different.

New test was using market instead of label